+++ internal/infra/database/migrations/20230221105536_create_eduprogscheme_table.up.sql
@@ -5,7 +5,7 @@ CREATE TABLE IF NOT EXISTS public.eduprogscheme(
     discipline_id integer ,
     eduprog_id integer REFERENCES eduprog(id) ON DELETE CASCADE,
     eduprogcomp_id integer REFERENCES eduprogcomp(id) ON DELETE CASCADE,
-    credits_per_semester integer,
+    credits_per_semester float8,
     created_date TIMESTAMP DEFAULT now(),
     updated_date TIMESTAMP DEFAULT now()
     );
